Refactoring + getting rid of linting problems. Code will still be broken throughouht the codebase since the dependency paths are now broken
