+++ .gitignore
@@ -1,3 +1,6 @@
+logs/
+.vscode/
+
 # Byte-compiled / optimized / DLL files
 __pycache__/
 *.py[cod]
